# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ЖерБақылау API",
    "description": "Цифровой мониторинг земель: панель инспектора, «Народный контроль» и статусы заявлений.\n\nОшибки возвращаются в формате `{error: {code, message, details}}`.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/audit/verify": {
      "get": {
        "tags": [
          "audit"
        ],
        "summary": "Verify the audit hash chain",
        "operationId": "verify_audit_api_v1_audit_verify_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChainCheck"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/stats/dashboard": {
      "get": {
        "tags": [
          "stats"
        ],
        "summary": "Dashboard",
        "operationId": "dashboard_api_v1_stats_dashboard_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardStats"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/satellite/ndvi": {
      "get": {
        "tags": [
          "satellite"
        ],
        "summary": "NDVI layer (Sentinel-2 L2A or demo provider)",
        "operationId": "ndvi_layer_api_v1_satellite_ndvi_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NdviLayer"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/satellite/scan": {
      "post": {
        "tags": [
          "satellite"
        ],
        "summary": "Start a satellite scan in the background (result arrives as a satellite.scan_completed event)",
        "operationId": "satellite_scan_api_v1_satellite_scan_post",
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SatelliteScanStarted"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/events": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Event feed (realtime fallback)",
        "operationId": "list_events_api_v1_events_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Return events with id > since",
              "title": "Since"
            },
            "description": "Return events with id > since"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventList"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/export/parcels.geojson": {
      "get": {
        "tags": [
          "export"
        ],
        "summary": "Export Geojson",
        "operationId": "export_geojson_api_v1_export_parcels_geojson_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/geo+json": {}
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/export/parcels.csv": {
      "get": {
        "tags": [
          "export"
        ],
        "summary": "Export Csv",
        "operationId": "export_csv_api_v1_export_parcels_csv_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/csv": {}
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChainCheck": {
        "properties": {
          "intact": {
            "type": "boolean",
            "title": "Intact"
          },
          "length": {
            "type": "integer",
            "title": "Length"
          },
          "head": {
            "type": "string",
            "title": "Head"
          },
          "broken_at_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Broken At Id"
          }
        },
        "type": "object",
        "required": [
          "intact",
          "length",
          "head"
        ],
        "title": "ChainCheck"
      },
...
      "DashboardStats": {
        "properties": {
          "kpi": {
            "$ref": "#/components/schemas/Kpi"
          },
          "signals_by_day": {
            "items": {
              "$ref": "#/components/schemas/DayCount"
            },
            "type": "array",
            "title": "Signals By Day"
          },
          "violations_by_type": {
            "items": {
              "$ref": "#/components/schemas/ViolationTypeCount"
            },
            "type": "array",
            "title": "Violations By Type"
          },
          "status_counts": {
            "items": {
              "$ref": "#/components/schemas/StatusCount"
            },
            "type": "array",
            "title": "Status Counts"
          },
          "upcoming_deadlines": {
            "items": {
              "$ref": "#/components/schemas/UpcomingDeadline"
            },
            "type": "array",
            "title": "Upcoming Deadlines"
          }
        },
        "type": "object",
        "required": [
          "kpi",
          "signals_by_day",
          "violations_by_type",
          "status_counts",
          "upcoming_deadlines"
        ],
        "title": "DashboardStats"
      },
      "DayCount": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "date",
          "count"
        ],
        "title": "DayCount"
      },
...
      "ErrorInfo": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorInfo"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorInfo"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
      "EventList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EventOut"
            },
            "type": "array",
            "title": "Items"
          },
          "last_id": {
            "type": "integer",
            "title": "Last Id"
          }
        },
        "type": "object",
        "required": [
          "items",
          "last_id"
        ],
        "title": "EventList"
      },
      "EventOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "payload",
          "created_at"
        ],
        "title": "EventOut"
      },
...
      "Kpi": {
        "properties": {
          "parcels_total": {
            "type": "integer",
            "title": "Parcels Total"
          },
          "area_total_ha": {
            "type": "number",
            "title": "Area Total Ha"
          },
          "violations_active": {
            "type": "integer",
            "title": "Violations Active"
          },
          "in_remediation": {
            "type": "integer",
            "title": "In Remediation"
          },
          "overdue": {
            "type": "integer",
            "title": "Overdue"
          },
          "signals_7d": {
            "type": "integer",
            "title": "Signals 7D"
          },
          "signals_new": {
            "type": "integer",
            "title": "Signals New"
          },
          "avg_reaction_hours": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Reaction Hours"
          }
        },
        "type": "object",
        "required": [
          "parcels_total",
          "area_total_ha",
          "violations_active",
          "in_remediation",
          "overdue",
          "signals_7d",
          "signals_new",
          "avg_reaction_hours"
        ],
        "title": "Kpi"
      },
...
      "NdviLayer": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "is_demo": {
            "type": "boolean",
            "title": "Is Demo"
          },
          "scanned_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scanned At"
          },
          "observed_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed From"
          },
          "observed_to": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed To"
          },
          "scenes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scenes"
          },
          "threshold": {
            "type": "number",
            "title": "Threshold"
          },
          "readings": {
            "items": {
              "$ref": "#/components/schemas/NdviReadingOut"
            },
            "type": "array",
            "title": "Readings"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "is_demo",
          "scanned_at",
          "threshold",
          "readings"
        ],
        "title": "NdviLayer"
      },
...
      "NdviReadingOut": {
        "properties": {
          "parcel_id": {
            "type": "string",
            "title": "Parcel Id"
          },
          "ndvi": {
            "type": "number",
            "title": "Ndvi"
          },
          "flagged": {
            "type": "boolean",
            "title": "Flagged"
          },
          "scanned_at": {
            "type": "string",
            "format": "date-time",
            "title": "Scanned At"
          },
          "observed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed At",
            "description": "Acquisition time of the satellite scene"
          },
          "scene_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scene Id"
          }
        },
        "type": "object",
        "required": [
          "parcel_id",
          "ndvi",
          "flagged",
          "scanned_at"
        ],
        "title": "NdviReadingOut"
      },
...
      "ParcelStatus": {
        "type": "string",
        "enum": [
          "OK",
          "UNDER_CHECK",
          "VIOLATION",
          "IN_REMEDIATION",
          "RESOLVED",
          "RETURNED_TO_STATE"
        ],
        "title": "ParcelStatus"
      },
...
      "SatelliteScanStarted": {
        "properties": {
          "started": {
            "type": "boolean",
            "title": "Started",
            "description": "False if a scan is already running"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "is_demo": {
            "type": "boolean",
            "title": "Is Demo"
          }
        },
        "type": "object",
        "required": [
          "started",
          "provider",
          "is_demo"
        ],
        "title": "SatelliteScanStarted"
      },
...
      "StatusCount": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/ParcelStatus"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "status",
          "count"
        ],
        "title": "StatusCount"
      },
...
      "UpcomingDeadline": {
        "properties": {
          "parcel_id": {
            "type": "string",
            "title": "Parcel Id"
          },
          "cadastral_number": {
            "type": "string",
            "title": "Cadastral Number"
          },
          "status": {
            "$ref": "#/components/schemas/ParcelStatus"
          },
          "violation_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ViolationType"
              },
              {
                "type": "null"
              }
            ]
          },
          "deadline_at": {
            "type": "string",
            "format": "date-time",
            "title": "Deadline At"
          },
          "days_left": {
            "type": "integer",
            "title": "Days Left"
          }
        },
        "type": "object",
        "required": [
          "parcel_id",
          "cadastral_number",
          "status",
          "violation_type",
          "deadline_at",
          "days_left"
        ],
        "title": "UpcomingDeadline"
      },
      "ViolationType": {
        "type": "string",
        "enum": [
          "UNUSED",
          "SELF_SEIZURE",
          "DUMP",
          "MISUSE"
        ],
        "title": "ViolationType"
      },
      "ViolationTypeCount": {
        "properties": {
          "violation_type": {
            "$ref": "#/components/schemas/ViolationType"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "violation_type",
          "count"
        ],
        "title": "ViolationTypeCount"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Supabase access token (or local dev token)",
        "scheme": "bearer"
      }
    }
  }
}